+++ packages/astro/test/preact-compat-component.test.ts
@@ -1,7 +1,7 @@
 import assert from 'node:assert/strict';
-import { after, before, describe, it } from 'node:test';
+import { before, describe, it } from 'node:test';
 import * as cheerio from 'cheerio';
-import { type DevServer, type Fixture, loadFixture } from './test-utils.ts';
+import { type Fixture, loadFixture } from './test-utils.ts';
 
 describe('Preact compat component', () => {
 	let fixture: Fixture;
@@ -10,26 +10,6 @@ describe('Preact compat component', () => {
 		fixture = await loadFixture({
 			root: './fixtures/preact-compat-component/',
 			outDir: './dist/preact-compat-component/',
-		});
-	});
-
-	describe('Development', () => {
-		let devServer: DevServer;
-
-		before(async () => {
-			devServer = await fixture.startDevServer();
-		});
-
-		after(async () => {
-			await devServer.stop();
-		});
-
-		it('Can load Counter', async () => {
-			const res = await fixture.fetch('/');
-			const html = await res.text();
-			const $ = cheerio.load(html);
-
-			assert.equal($('#counter-text').text(), '0');
 		});
 	});