+++ packages/astro/test/static-build-page-dist-url.test.ts
@@ -18,6 +18,7 @@ describe('Static build: pages routes have distURL', () => {
 				},
 			],
 			outDir: './dist/static-build-page-dist-url/',
+			cacheDir: './node_modules/.astro-test/static-build-page-dist-url/',
 		});
 		await fixture.build();
 	});